Module connect6::board

Connect6 boards.

Structs

Board

A Connect6 board.

ParsePointError

An error which can be returned when parsing a point.

Point

A 2D point with u32 coordinates.

Slot

A slot (namely intersection) on the board.

Enums

Axis

Axes on the board.

Stone

A stone on the board, either black or white.