Struct connect6::board::Board

pub struct Board { /* private fields */ }

A Connect6 board.

Implementations

impl Board

pub fn new(size: u32) -> Board

Creates a new Board with the given size.

Panics

Panics if the size is zero, larger than 255 or is even.

pub fn size(&self) -> u32

Returns the size of the board.

pub fn get(&self, p: Point) -> Option<&Slot>

Returns a reference to a slot, or None if the point is out of board.

pub fn get_mut(&mut self, p: Point) -> Option<&mut Slot>

Returns a mutable reference to a slot, or None if the point is out of board.

pub fn contains_point(&self, p: Point) -> bool

Returns true if the board contains a point.

pub fn move_index(&self) -> u32

Returns the current move index starting from 1, or 0 if the board is empty.

pub fn is_full(&self) -> bool

Returns true if the board is full.

pub fn make_move(&mut self, mov: (Point, Point), stone: Stone)

Makes a move on the board.

Panics

Panics when moving out of board or into an occupied slot.

pub fn check_win(&self, p: Point, stone: Stone) -> bool

Returns true if there is a six or overline of the given stone through the point.

This method assumes that the slot at the point is occupied by the given stone.